+++ NEWS.md
@@ -1,3 +1,9 @@
+# rasciidoc 0.2.0
+
+* Added function to adjust knitr's hooks for asciidoc files, providing a work
+  around for not using asciidoc's [MESSAGE|WARNING|ERROR] when knitting produces
+  a message|warning|error.
+
 # rasciidoc 0.1.0
 
 * Added core functionality.